+++ public/src/core/utils/appUtils.ts
@@ -1,27 +1,4 @@
-import { SpatialObjectCodeEnum } from '@api-client';
 import moment = require('moment');
-
-export const DELIMITER = {
-  PIPE: '|'
-}
-
-export const SpatialTypeMap = new Map<SpatialObjectCodeEnum, object>([
-  [SpatialObjectCodeEnum.CutBlock, {
-    source: 'cut_block',
-    type: 'Polygon',
-    desc: 'Cut Block'
-  }],
-  [SpatialObjectCodeEnum.RoadSection, {
-    source: 'road_section',
-    type: 'LineString',
-    desc: 'Road Section'
-  }],
-  [SpatialObjectCodeEnum.Wtra, {
-    source: 'retention_area',
-    type: 'Polygon',
-    desc: 'Retention Area'
-  }],
-]);
 
 /**
  * @class AppUtils
